requirements.txt
# Framework principal
fastapi
uvicorn[standard]

# ORM e banco
sqlalchemy
psycopg2-binary
alembic
requests>=2.31.0
apscheduler>=3.10.4


# Validação e segurança
pydantic[email]
email-validator
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]
python-dotenv

# Qualidade de código
black
autopep8
flake8
isort
mypy

